refactor: move `AddPhotoCard` to the `features` folder
